% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{createAOI}
\alias{createAOI}
\title{Create Area of Interest (AOI)}
\usage{
createAOI(aoi_name, description, geoJSON)
}
\arguments{
\item{aoi_name}{The name of the Area of Interest (AOI).}

\item{description}{A description of the AOI.}

\item{geoJSON}{The geoJSON defining the AOI.}
}
\value{
The response from the API call, typically containing details about the created AOI.
}
\description{
This function creates an Area of Interest (AOI) by making an API call with the provided parameters.
}
\examples{
\dontrun{
  response <- createAOI(aoi_name = "My AOI", description = "Description of my AOI", geoJSON = geojson_data)
  print(response)
}
}
