% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getLayerNumbers}
\alias{getLayerNumbers}
\title{Get Layer Numbers}
\usage{
getLayerNumbers()
}
\value{
A named vector where the names are descriptions of the layers and the values are their corresponding numbers.
}
\description{
This function returns a named vector of layer numbers and their corresponding descriptions.
}
\examples{
layers <- getLayerNumbers()
print(layers)
}
