% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getSingleLayer}
\alias{getSingleLayer}
\title{Get Single Layer}
\usage{
getSingleLayer(cell_code, aoi_id, layer)
}
\arguments{
\item{cell_code}{The cell code for the layer to be downloaded.}

\item{aoi_id}{The Area of Interest (AOI) ID for the layer to be downloaded.}

\item{layer}{The layer number to be downloaded.}
}
\value{
A character vector containing the file paths of the unzipped files.
}
\description{
This function downloads a single layer for a specified cell code and Area of Interest (AOI) ID, unzips the downloaded file, and returns the list of files.
}
\examples{
\dontrun{
  files <- getSingleLayer(cell_code = "12345", aoi_id = "67890", layer = "1")
  print(files)
}
}
