% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{listAOIProducts}
\alias{listAOIProducts}
\title{List AOI Products}
\usage{
listAOIProducts(
  aoi_id,
  start_date = NULL,
  end_date = NULL,
  all = F,
  verbose = F
)
}
\arguments{
\item{aoi_id}{The ID of the Area of Interest (AOI).}

\item{start_date}{The start date for the product search (optional, default is NULL).}

\item{end_date}{The end date for the product search (optional, default is NULL).}

\item{all}{Logical, if TRUE, retrieves all products regardless of date (default is FALSE).}

\item{verbose}{Logical, if TRUE, prints additional information during the function execution (default is FALSE).}
}
\value{
An \code{sf} object containing the products for the specified AOI.
}
\description{
This function retrieves the list of products for a specified Area of Interest (AOI) and converts them into an \code{sf} object.
}
\note{
By default, the function retrieves the latest product for the specified AOI. If the \code{start_date} and \code{end_date} parameters are provided, the function retrieves products within the specified date range.
}
\examples{
\dontrun{
  products <- listAOIProducts(aoi_id = "12345", start_date = "2023-01-01", end_date = "2023-01-31")
  print(products)
}
}
