% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{listUserAOI}
\alias{listUserAOI}
\title{List User AOIs}
\usage{
listUserAOI()
}
\value{
An \code{sf} object containing the user's AOIs.
}
\description{
This function retrieves the list of Areas of Interest (AOIs) for the current user and converts them into an \code{sf} object.
}
\examples{
\dontrun{
  user_aois <- listUserAOI()
  print(user_aois)
}
}
