% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{orderMaxFloodExtent}
\alias{orderMaxFloodExtent}
\title{Order Maximum Flood Extent}
\usage{
orderMaxFloodExtent(aoi_id, start_date, end_date, output_type = "both")
}
\arguments{
\item{aoi_id}{The Area of Interest (AOI) ID.}

\item{start_date}{The start date for the flood extent order.}

\item{end_date}{The end date for the flood extent order.}

\item{output_type}{The type of output to be returned. Must be one of "both", "raster", or "vector". Default is "both".}
}
\value{
The response from the API call, typically containing details about the order.
}
\description{
This function orders the maximum flood extent for a specified Area of Interest (AOI) ID and date range.
}
\examples{
\dontrun{
  response <- orderMaxFloodExtent(aoi_id = "12345", start_date = "2023-01-01", end_date = "2023-01-31", output_type = "both")
  print(response)
}
}
